+++ bootstrap_zk.bash
@@ -11,16 +11,14 @@ declare -r ARCHIVE='apache-zookeeper-'"$ZK_VERSION"'-bin.'"${EXT}"
 mkdir -p "$DOWNLOAD_DIR" "$INSTALL_DIR"
 pushd "$DOWNLOAD_DIR"
 
-curl -L 'https://archive.apache.org/dist/zookeeper/zookeeper-'"$ZK_VERSION"'/'"$ARCHIVE" -o "$ARCHIVE"
+[ -s "$ARCHIVE" ] || curl -L 'https://archive.apache.org/dist/zookeeper/zookeeper-'"$ZK_VERSION"'/'"$ARCHIVE" -o "$ARCHIVE"
 tar xzf "$ARCHIVE" -C "$INSTALL_DIR" --strip-components=1
 
 popd
 pushd "$INSTALL_DIR"
 
-ls -alR "$HOME/zk"
-
 echo "tickTime=2000
-dataDir=$HOME/zk/data_
-clientPort=2181" > "$HOME"'/zk/conf/zoo.cfg'
+dataDir=$INSTALL_DIR/data_
+clientPort=2181" > "$INSTALL_DIR"'/conf/zoo.cfg'
 
 bin/zkServer.sh start
